requirements.txt
aiogram==3.4.1
aiohttp==3.9.3
python-dotenv==1.0.1
apscheduler
pytz
